refactored queries and api to make use of bluebird promises. modified readme.md
