+++ package.json
@@ -18,6 +18,7 @@
     "babel-jest": "^25.2.4",
     "babel-plugin-transform-export-extensions": "^6.22.0",
     "babel-polyfill": "^6.26.0",
+    "bluebird": "^3.7.2",
     "cors": "^2.8.5",
     "css-loader": "^3.4.2",
     "express": "^4.17.1",
